#25: Re-factored ItemMixin and AttributeMixin to inherit from a parent class named IndexOrAccessToChainMixin that provides the core functionality of this feature.
