+++ rightshift/chains.py
@@ -97,3 +97,27 @@ default = Default
 """
 TODO: Document
 """
+
+
+class IndexOrAccessToChainMixin(object):
+    """
+
+    """
+    __chain_class = None
+    __class = None
+
+    def __getattr__(self, item_name):
+        """
+        :param item_name: A valid item name value
+        :return: an ItemChain instance
+        :rtype: ItemChain
+        """
+        return self.__chain_class(self, self.__class(item_name))
+
+    def __getitem__(self, item_or_slice):
+        """
+        :param item_or_slice: A valid item name or slice value
+        :return: an ItemChain instance
+        :rtype: ItemChain
+        """
+        return self.__chain_class(self, self.__class(item_or_slice))
